requirements.txt
uvicorn==0.41.0
fastapi==0.135.1
psycopg2-binary==2.9.11
